Move code for running SQL statements into a separate class
This is only done for better readability and for easing future changes.
No major logic changes intended here.
