+++ CMakeLists.txt
@@ -144,6 +144,7 @@ set(SQLB_MOC_HDR
 	src/DotenvFormat.h
 	src/Palette.h
 	src/CondFormat.h
+	src/RunSql.h
 )
 
 set(SQLB_SRC
@@ -194,6 +195,7 @@ set(SQLB_SRC
 	src/DotenvFormat.cpp
 	src/Palette.cpp
 	src/CondFormat.cpp
+	src/RunSql.cpp
 )
 
 set(SQLB_FORMS
